+++ backend/package.json
@@ -27,13 +27,12 @@
     "underscore": "^1.9.1",
     "winston": "^3.2.1",
     "winston-daily-rotate-file": "^4.2.1",
-    "youch": "^2.0.10",
-    "yup": "^0.27.0"
+    "youch": "^2.0.10"
   },
   "devDependencies": {
     "@babel/core": "^7.6.4",
     "@babel/preset-env": "^7.6.3",
-    "eslint": "^6.7.1",
+    "eslint": "^6.7.2",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-config-prettier": "^6.5.0",
     "eslint-config-standard": "^14.1.0",
