+++ backend/src/app/controllers/CriaProcessoController.js
@@ -3,7 +3,6 @@
 /* eslint-disable camelcase */
 import VDadosPessoa from '../models/VDadosPessoa';
 import Processo from '../models/Processo';
-import ProcessoValidator from '../validators/ProcessoValidator';
 import AuditoriaController from './AuditoriaController';
 
 class CriaProcessoController {
@@ -20,10 +19,6 @@ class CriaProcessoController {
     }
 
     async store(req, res) {
-        const validator = new ProcessoValidator();
-        if (!(await validator.validate(req))) {
-            return res.status(400).json({ error: validator.errors });
-        }
         if (req.body.pro_nome === '') {
             req.body.pro_nome = null;
         }
