+++ backend/src/app/controllers/RazaoTramiteController.js
@@ -1,60 +1,60 @@
 /* eslint-disable consistent-return */
 /* eslint-disable func-names */
 /* eslint-disable camelcase */
-import Genero from '../models/Genero';
+import RazaoTramite from '../models/RazaoTramite';
 import AuditoriaController from './AuditoriaController';
 
-class GeneroController {
+class RazaoTramiteController {
     async index(req, res) {
-        const generos = await Genero.findAll({
-            order: ['gen_nome'],
-            attributes: ['gen_id', 'gen_nome'],
+        const razoes = await RazaoTramite.findAll({
+            order: ['raz_nome'],
+            attributes: ['raz_id', 'raz_nome'],
             logging: false
         });
-        return res.json(generos);
+        return res.json(razoes);
     }
 
     async store(req, res) {
-        const { gen_id, gen_nome } = await Genero.create(req.body, {
+        const { raz_id, raz_nome } = await RazaoTramite.create(req.body, {
             logging: false
         });
         // auditoria de inserção
-        AuditoriaController.audita(req.body, req, 'I', gen_id);
+        AuditoriaController.audita(req.body, req, 'I', raz_id);
         //
         return res.json({
-            gen_id,
-            gen_nome
+            raz_id,
+            raz_nome
         });
     }
 
     async update(req, res) {
-        const genero = await Genero.findByPk(req.params.id, { logging: false });
+        const razao = await RazaoTramite.findByPk(req.params.id, { logging: false });
         // auditoria de edição
         AuditoriaController.audita(
-            genero._previousDataValues,
+            razao._previousDataValues,
             req,
             'U',
             req.params.id
         );
         //
-        if (!genero) {
-            return res.status(400).json({ error: 'Gênero não encontrado' });
+        if (!razao) {
+            return res.status(400).json({ error: 'Razão de trâmite não encontrado' });
         }
-        await genero.update(req.body, { logging: false });
-        return res.json(genero);
+        await razao.update(req.body, { logging: false });
+        return res.json(razao);
     }
 
     async delete(req, res) {
-        const genero = await Genero.findByPk(req.params.id, { logging: false });
-        if (!genero) {
-            return res.status(400).json({ error: 'Gênero não encontrado' });
+        const razao = await RazaoTramite.findByPk(req.params.id, { logging: false });
+        if (!razao) {
+            return res.status(400).json({ error: 'Razão de trâmite não encontrado' });
         }
-        await genero
+        await razao
             .destroy({ logging: false })
             .then(auditoria => {
                 // auditoria de deleção
                 AuditoriaController.audita(
-                    genero._previousDataValues,
+                    razao._previousDataValues,
                     req,
                     'D',
                     req.params.id
@@ -64,11 +64,11 @@ class GeneroController {
             .catch(function(err) {
                 if (err.toString().includes('SequelizeForeignKeyConstraintError')) {
                     return res.status(400).json({
-                        error: 'Erro ao excluir gênero. O gênero possui uma ou mais ligações.'
+                        error: 'Erro ao excluir razão de trâmite. A razão de trâmite possui uma ou mais ligações.'
                     });
                 }
             });
         return res.send();
     }
 }
-export default new GeneroController();
+export default new RazaoTramiteController();
