+++ backend/src/app/controllers/Spa2Controller.js
@@ -1,7 +1,7 @@
-class BibController {
+class Spa2Controller {
     async index(req, res) {
         return res.json('spa2');
     }
 }
 
-export default new BibController();
+export default new Spa2Controller();
