--- backend/src/app/validators/ModeloMenuValidator.js
@@ -1,26 +1,0 @@
-import * as Yup from 'yup';
-
-class ModeloMenuValidator {
-    constructor() {
-        this.init();
-        this.errors = [];
-    }
-
-    init() {
-        this.schema = Yup.object().shape({
-            mmu_nome: Yup.string().required()
-        });
-    }
-
-    async validate(req) {
-        try {
-            await this.schema.validate(req.body, { abortEarly: false });
-        } catch (e) {
-            this.errors = e.errors;
-            return false;
-        }
-        return true;
-    }
-}
-
-export default ModeloMenuValidator;
